feat (view) : Input / Output 구현
기존 GameView 하나에서 Input/Output으로 기능을 나눔
